chore: move get_version to ckb executable crate
Avoid abusing it, since the version is only meaningful when used in the
executable crate.
