+++ util/build-info/src/lib.rs
@@ -1,41 +1,3 @@
-use std::env;
-
-#[macro_export]
-macro_rules! get_version {
-    () => {{
-        let major = env!("CARGO_PKG_VERSION_MAJOR")
-            .parse::<u8>()
-            .expect("CARGO_PKG_VERSION_MAJOR parse success");
-        let minor = env!("CARGO_PKG_VERSION_MINOR")
-            .parse::<u8>()
-            .expect("CARGO_PKG_VERSION_MINOR parse success");
-        let patch = env!("CARGO_PKG_VERSION_PATCH")
-            .parse::<u16>()
-            .expect("CARGO_PKG_VERSION_PATCH parse success");
-        let dash_pre = {
-            let pre = env!("CARGO_PKG_VERSION_PRE");
-            if pre == "" {
-                pre.to_string()
-            } else {
-                "-".to_string() + pre
-            }
-        };
-
-        let host_compiler = $crate::get_channel();
-        let commit_describe = option_env!("COMMIT_DESCRIBE").map(ToString::to_string);
-        let commit_date = option_env!("COMMIT_DATE").map(ToString::to_string);
-        Version {
-            major,
-            minor,
-            patch,
-            dash_pre,
-            host_compiler,
-            commit_describe,
-            commit_date,
-        }
-    }};
-}
-
 // some code taken and adapted from RLS and cargo
 #[derive(Debug, Default, Clone)]
 pub struct Version {
@@ -43,7 +5,6 @@ pub struct Version {
     pub minor: u8,
     pub patch: u16,
     pub dash_pre: String,
-    pub host_compiler: Option<String>,
     pub commit_describe: Option<String>,
     pub commit_date: Option<String>,
 }
@@ -91,15 +52,6 @@ impl std::fmt::Display for Version {
         }
 
         Ok(())
-    }
-}
-
-pub fn get_channel() -> Option<String> {
-    if let Ok(channel) = env::var("CFG_RELEASE_CHANNEL") {
-        Some(channel)
-    } else {
-        // we could ask ${RUSTC} -Vv and do some parsing and find out
-        Some(String::from("nightly"))
     }
 }
